+++ srcs/pf_buffer.c
@@ -6,7 +6,7 @@
 /*   By: angavrel <angavrel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/05/06 06:02:27 by angavrel          #+#    #+#             */
-/*   Updated: 2017/05/06 06:02:29 by angavrel         ###   ########.fr       */
+/*   Updated: 2017/05/06 15:57:34 by angavrel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,16 +18,16 @@ void		buffer(t_printf *p, void *new, size_t size)
 	long long	new_i;
 
 	new_i = 0;
-	while (PF_BUF_SIZE - p->buffer_index < (int)size)
+	while (PF_BUF_SIZE - p->buffer_index < size)
 	{
-		diff = PF_BUF_SIZE - p->buffer_index;
-		ft_memcpy(&(p->buff[p->buffer_index]), &(new[new_i]), diff);
-		size -= diff;
-		new_i += diff;
-		p->buffer_index += diff;
-		p->len += diff;
-		write(p->fd, p->buff, p->buffer_index);
-		p->buffer_index = 0;
+			diff = PF_BUF_SIZE - p->buffer_index;
+			ft_memcpy(&(p->buff[p->buffer_index]), &(new[new_i]), diff);
+			size -= diff;
+			new_i += diff;
+			p->buffer_index += diff;
+			p->len += diff;
+			write(p->fd, p->buff, p->buffer_index);
+			p->buffer_index = 0;
 	}
 	ft_memcpy(&(p->buff[p->buffer_index]), &(new[new_i]), size);
 	p->buffer_index += size;
